refactor: move the presentation out of the core
