+++ core/CVApplication.js
@@ -1,5 +1,6 @@
 import { DataLoader } from './DataLoader.js';
 import { RendererContainer } from './RendererContainer.js';
+import { ErrorRenderer } from '../renderers/ErrorRenderer.js';
 
 /**
  * Main CV Application Controller
@@ -7,11 +8,12 @@ import { RendererContainer } from './RendererContainer.js';
  */
 export class CVApplication {
   constructor(dataLoader = new DataLoader(), rendererContainer = new RendererContainer(),
-              documentLocalizer = null, i18n = null) {
+              documentLocalizer = null, i18n = null, errorRenderer = new ErrorRenderer()) {
     this.dataLoader = dataLoader;
     this.rendererContainer = rendererContainer;
     this.documentLocalizer = documentLocalizer;
     this.i18n = i18n;
+    this.errorRenderer = errorRenderer;
     this.isInitialized = false;
   }
 
@@ -49,17 +51,11 @@ export class CVApplication {
    */
   handleError(root, error) {
     console.error('CV Application Error:', error);
-    const title = this.i18n ? this.i18n.t('errors.loadingTitle') : 'Error loading CV';
-    const hint = this.i18n
-      ? this.i18n.t('errors.loadingHint')
-      : 'Please check the console for more details.';
-    root.body.innerHTML = `
-      <div style="text-align: center; padding: 50px; color: #666;">
-        <h2>${title}</h2>
-        <p>${error.message}</p>
-        <p>${hint}</p>
-      </div>
-    `;
+    this.errorRenderer.render(root, {
+      title: this.i18n ? this.i18n.t('errors.loadingTitle') : 'Error loading CV',
+      message: error.message,
+      hint: this.i18n ? this.i18n.t('errors.loadingHint') : 'Please check the console for more details.'
+    });
   }
 
   /**
